biopython_rewrite/requirements.txt
biopython==1.82
pandas
numpy
pytest
matplotlib
seaborn
snakemake==7.26

